+++ KnightTournamentWithRoute/KnightTournament/Models/CombatsKnight.cs
@@ -6,7 +6,7 @@ namespace KnightTournament.Models
     {
         public Guid Id { get; set; } = Guid.NewGuid();
 
-        public double Points { get; set; } = 0;
+        public int Points { get; set; } = 0;
 
         public virtual Guid? CombatId { get; set; }
